Fix a bug where moves were moves were incorrectly disambiguated.
- In some cases, where two pieces could move to the same square, one move would get removed as it would
put the king in check. The removal of this move removed the disambiguation condition from the other.
When the pieces were found, in some rare instances the piece that would put the king in check
would try to be moved instead, even though technically this move was removed.
This would cause an illegal move error as you're not allowed to move such that
your king is put in check.