Move checking end states for check inside short_algebraic_to_long.
- The rationale behind this comes from the Ethereal vs Lc0 game from TCEC Season 14.
- In the PGN, a rook move (Rg2+) is left ambiguous, because the rook on g7 cannot move, since doing so
would place black in check. However, my code would find this rook first, since the code is not disambiguated,
see that it can move to g2, and then move it. Then it would go to check move legality, where it would find out
that moving it puts you in check, and then return False, saying it wasn't a legal move. However, the code
should recognize that yes, this rook is illegal, however the rook on g1 is legal to move to g2.
- Code now checks that if it leaves you in check in short_algebraic, and if it does, looks to see if there's another
piece that could possibly make this move. This is marginally slower as more boolean comparisons will be done
per loop, but in general the same amount of pieces are checked every time and the time added is negligable.