Reorder root node moves every odd ply.
- Used to only do it after ply 1, but it would search moves that it didn't know
where a checkmate sequence at ply 1 first in some cases. This is a problem as if the search
gets cutoff before the reaches the non checkmate sequences it'll play into a checkmate.
By reordering every odd ply we now put known non-checkmate sequences first. Guaranteeing that
no matter where the search gets cutoff it'll be a non-checkmate sequence.